test: TravelPlanRequestBuilder와 PlanDayRequestBuilder 별도 클래스로 분리
